feat: add relationship checks before deletion in Delete controllers

+++ src/Controller/Subscription/DeleteSubscriptionPlanController.php
@@ -3,6 +3,8 @@
 namespace App\Controller\Subscription;
 
 use App\Entity\SubscriptionPlan;
+use App\Enum\ApiError;
+use App\Trait\ApiResponseTrait;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -11,6 +13,8 @@ use Symfony\Component\HttpKernel\Attribute\AsController;
 #[AsController]
 class DeleteSubscriptionPlanController extends AbstractController
 {
+    use ApiResponseTrait;
+
     public function __construct(
         private EntityManagerInterface $entityManager
     ) {
@@ -18,9 +22,16 @@ class DeleteSubscriptionPlanController extends AbstractController
 
     public function __invoke(SubscriptionPlan $data): JsonResponse
     {
+        // Vérifier si le plan a des abonnements actifs
+        if ($data->getSubscriptions()->count() > 0) {
+            return $this->errorResponse(409, ApiError::SUBSCRIPTION_PLAN_HAS_SUBSCRIPTIONS, [
+                'subscriptionsCount' => $data->getSubscriptions()->count()
+            ]);
+        }
+
         $this->entityManager->remove($data);
         $this->entityManager->flush();
 
-        return $this->json(['message' => 'Plan d\'abonnement supprimé avec succès'], 204);
+        return $this->json(['code' => 204, 'data' => null], 204);
     }
 }
