+++ src/Ecommerce/Controller/Product/DeleteProductController.php
@@ -3,6 +3,8 @@
 namespace App\Ecommerce\Controller\Product;
 
 use App\Ecommerce\Entity\Product;
+use App\Enum\ApiError;
+use App\Trait\ApiResponseTrait;
 use Doctrine\ORM\EntityManagerInterface;
 use Symfony\Bundle\FrameworkBundle\Controller\AbstractController;
 use Symfony\Component\HttpFoundation\JsonResponse;
@@ -11,6 +13,8 @@ use Symfony\Component\HttpKernel\Attribute\AsController;
 #[AsController]
 class DeleteProductController extends AbstractController
 {
+    use ApiResponseTrait;
+
     public function __construct(
         private EntityManagerInterface $entityManager
     ) {
@@ -18,9 +22,16 @@ class DeleteProductController extends AbstractController
 
     public function __invoke(Product $data): JsonResponse
     {
+        // Vérifier si le produit est dans des commandes
+        if ($data->getOrderItems()->count() > 0) {
+            return $this->errorResponse(409, ApiError::PRODUCT_HAS_ORDERS, [
+                'ordersCount' => $data->getOrderItems()->count()
+            ]);
+        }
+
         $this->entityManager->remove($data);
         $this->entityManager->flush();
 
-        return $this->json(['message' => 'Produit supprimé avec succès'], 204);
+        return $this->json(['code' => 204, 'data' => null], 204);
     }
 }
